refactor scheduling: create scheduler directory..
